feat: 디스코드 연동하기 API 구현 (#98)
* refactor: 응답 필드 설명 수정

* feat: 디스코드 유저네임 업데이트 시 중복검사 로직 추가

* feat: 디스코드 연동하기 API 구현

* feat: 디스코드 회원정보 수정 허용하지 않도록 변경

* feat: 디스코드 연동하기 서비스 로직 구현

* feat: 디스코드 회원정보 수정 서비스 메서드 비활성화 처리

* refactor: 인증코드에 Integer 사용하도록 변경

* fix: 요청 바디로 받도록 수정

* feat: 디스코드 닉네임를 요청으로 받도록 추가

* fix: 멤버 업데이트 시 닉네임 전달하도록 수정

* fix: 객체 비교 방식 수정

* feat: 닉네임 및 유저네임 중복 검증 추가

* refactor: 코드 일치 확인 메서드를 캡슐화

* refactor: 상태 업데이트 메서드의 이름을 적절하게 수정

* feat: 레포지터리 메서드 추가

* feat: 에러코드 추가

* refactor: 에러코드를 static import 하도록 변경

* refactor: `@Range` 를 사용하도록 개선

+++ src/main/java/com/gdschongik/gdsc/domain/discord/dto/response/DiscordVerificationCodeResponse.java
@@ -2,9 +2,9 @@ package com.gdschongik.gdsc.domain.discord.dto.response;
 
 import java.time.Duration;
 
-public record DiscordVerificationCodeResponse(Long code, Duration ttl) {
+public record DiscordVerificationCodeResponse(Integer code, Duration ttl) {
 
-    public static DiscordVerificationCodeResponse of(Long code, Long ttlSeconds) {
+    public static DiscordVerificationCodeResponse of(Integer code, Long ttlSeconds) {
         return new DiscordVerificationCodeResponse(code, Duration.ofSeconds(ttlSeconds));
     }
 }
